+++ Parse/Config.hpp
@@ -1,3 +1,6 @@
+#ifndef CONFIG_HPP
+# define CONFIG_HPP
+
 #include <vector>
 #include <string>
 #include <map>
@@ -9,29 +12,19 @@
 #include <iostream>
 
 class ServerConfig;
-class LocationConfig;
 
 class Config
 {
 private:
-    std::map<std::string, ServerConfig> _servers;
 public:
-    Config(/* args */);
+    std::vector<ServerConfig> _servers;
     Config(int argc, char **argv);
+
     ~Config();
     void setServers(const std::map<std::string, ServerConfig> servers);
     std::map<std::string, ServerConfig> getServers();
+
+    void    parseConfig(const std::string filename);
 };
 
-Config::Config(/* args */)
-{
-}
-
-void    parseConfig(const std::string &filename)
-{
-    // parseServer(filename);
-}
-
-Config::~Config()
-{
-}
+#endif
