+++ Parse/ServerConfig.cpp
@@ -1,36 +1,17 @@
 #include "ServerConfig.hpp"
 
-class ServerConfig
-{
-public:
-    int _port;
-    std::string _server_name;
-    int _client_max_body_size;
-    std::unordered_map<int, std::string> _error_pages;
-    std::map<std::string, LocationConfig> _locations;
-    ServerConfig(/* args */);
-    ~ServerConfig();
-    friend Config;
-};
-
-ServerConfig::ServerConfig(/* args */)
-{
-}
-
 ServerConfig::~ServerConfig()
 {
 }
-
 
 void replaceTabsWithSpaces(std::string& str) {
     std::replace(str.begin(), str.end(), '\t', ' ');  // replace all tabs with spaces
 }
 
-ServerConfig    parseServer(const std::string &filename)
+ServerConfig    ServerConfig::parseServer(std::ifstream &file)
 {
-    std::ifstream   file(filename.c_str());
     std::string     line;
-    ServerConfig    currentConfig;
+    ServerConfig    currentServer;
     bool inServerBlock = false;
 
     while (getline(file, line)) {
@@ -39,40 +20,37 @@ ServerConfig    parseServer(const std::string &filename)
         std::string key;
         if (!(iss >> key)) continue; // Skip empty lines
 
+        if (line[0] ==  '}')
+        {
+            break ;
+        }
         if (key == "server") {
-            if (inServerBlock) {
-                currentConfig = ServerConfig(); // Reset
-            }
             inServerBlock = true;
         } else if (key == "listen") {
             std::string listen;
             iss >> listen;
-            currentConfig._port = std::stoi(listen);
+            currentServer._port = std::stoi(listen);
         } else if (key == "server_name") {
             std::string server_name;
             iss >> server_name;
-            currentConfig._server_name = server_name;
+            currentServer._server_name = server_name;
         } else if (key == "error_pages") {
-            currentConfig._error_pages = parseErrorPages(iss); // 발생 가능할 에러 생각
+            currentServer._error_pages = parseErrorPages(iss); // 발생 가능할 에러 생각
         } else if (key == "client_max_body_size") {
             std::string client_max_body_size;
             iss >> client_max_body_size;
-            currentConfig._client_max_body_size = std::stoi(client_max_body_size);
+            currentServer._client_max_body_size = std::stoi(client_max_body_size);
         } 
-        // else if (key == "location") {
-        //     LocationConfig location = parseLocation(iss);
-        //     currentConfig._locations[location._path] = location; }
-        else if (key == "}") {
-            inServerBlock = false;
-        }
-        if (key == "}") {
-            inServerBlock = false;
+        else if (key == "location") {
+            file.seekg(-(line.length() + 1), std::ios::cur);
+            LocationConfig location = LocationConfig::parseLocation(file);
+            currentServer._locations.insert(std::make_pair(location.getPath(), location));
         }
     }
-    return currentConfig;
+    return currentServer;
 }
 
-std::unordered_map<int, std::string> parseErrorPages(std::istringstream &iss) {
+std::unordered_map<int, std::string> ServerConfig::parseErrorPages(std::istringstream &iss) {
     std::unordered_map<int, std::string> error_pages;
     std::string error_code;
     std::string error_page;
